requirements.txt
beautifulsoup4==4.12.3
numpy==1.26.3
pandas==2.1.4
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
selenium==4.16.0
tornado==6.4
tornado-swagger==1.5.0